+++ CHANGELOG.md
@@ -1,11 +1,15 @@
+## 1.7.0
+
+- Improve `store.getState()` API (getState without arguments now equal `getStateInternal`)
+
 ## 1.6.0
 
-Add lazy reducers to createStore
+- Add lazy reducers to createStore
 
 ## 1.5.0
 
-Rewrite core - improve bundle size and performance
+- Rewrite core - improve bundle size and performance
 
 ## 1.4.0
 
-Add `createStore` - a half compatible **redux**-like API
+- Add `createStore` - a half compatible **redux**-like API